backend/pyproject.toml
[project]
name = "template-stack-backend"
version = "0.1.0"
description = "Template stack backend — Clerk + FastAPI + Neon/Postgres"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.9",
    "PyJWT[crypto]>=2.9.0",
    "anthropic>=0.39.0",
    "litellm>=1.50.0",
    "inngest>=0.4.0",
    "upstash-redis>=1.1.0",
    "sentry-sdk[fastapi]>=2.18.0",
    "langsmith>=0.1.140",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
